Update Features and Bug Fix
- Minor Fixed One Word Story
- Added Mine Command for Fun
- Changing Poll System
- Giveaway Command Become The Choosen One Command

BackSide :
- Update News
- Remove some Functions handler and send it into cog (not done yet)

+++ GamePack/ows.py
@@ -13,14 +13,42 @@ class OWS(commands.Cog):
         self.bot = bot
         self.conn = dbm.connect_db("./DataPack/ows_game.db")
 
+    def check_in_game_command(self, channel_id):
+        def inner_check(message):
+            if channel_id == message.channel.id:
+                return True
+            else:
+                return False
+        return inner_check
+
+    def check_last_person_ows_give_title(self, author):
+        def inner_check(message):
+            if author.id == message.author.id:
+                return True
+            else:
+                False
+        return inner_check
+
+    def check_choosen_story(self, author, list_story: list):
+        def inner_check(message):
+            try:
+                if message.author == author and (int(message.content) > 0 and int(message.content) <= len(list_story)):
+                    return True
+                else:
+                    return False
+            except Exception as exc:
+                print(type(exc), exc)
+                return False
+        return inner_check
+
     @commands.command(aliases=["OWS", "OWs", "Ows", "OwS", "oWS", "oWs", "owS"])
     async def ows(self, ctx, stat: str, story_num = None):
         statuses = ["start", "h", "help", "rdm", "random", "ourstory", "os", "how", "delete", "del", "read", "r"]
         try:
-            if stat.lower() not in statuses:
+            if stat.lower() not in statuses: # None of these known on this Command
                 raise commands.BadArgument
 
-            if stat.lower() == "help" or stat.lower() == "h": # Help
+            if stat.lower() == "help" or stat.lower() == "h": # Need Help about OWS
                 emb = discord.Embed(colour=discord.Colour(WHITE))
                 emb.set_author(name="📖 One Word Story")
                 emb.add_field(name="Commands (alias):", value=open("./DataPack/Help/owsh.txt", 'r').read(), inline=False)
@@ -38,11 +66,11 @@ class OWS(commands.Cog):
             if stat.lower() == "ourstory" or stat.lower() == "os": # Own Server Story to Look Up
                 await self.show_list_story(ctx)
             
-            if stat.lower() == "read" or stat.lower() == "r":
+            if stat.lower() == "read" or stat.lower() == "r": # Choose and Read server own story
                 list_svr_story, temp_msg = await self.show_list_story(ctx)
                 try:
                     xmsg = await ctx.send("***Pick a Story you want to read By Number.***")
-                    temp_msg1 = await self.bot.wait_for(event="message", check=check_choosen_story(ctx.message.author, list_svr_story), timeout=60.0)
+                    temp_msg1 = await self.bot.wait_for(event="message", check=self.check_choosen_story(ctx.message.author, list_svr_story), timeout=60.0)
                     picked_story = list_svr_story[int(temp_msg1.content) - 1]
                     emb_story = discord.Embed(title=picked_story[2], description=picked_story[3], colour=discord.Colour(WHITE))
                     await ctx.send(embed=emb_story)
@@ -60,7 +88,7 @@ class OWS(commands.Cog):
                 list_svr_story, temp_msg = await self.show_list_story(ctx)
                 try:
                     xmsg = await ctx.send("***Pick a Story you want to read By Number.***")
-                    temp_msg1 = await self.bot.wait_for(event="message", check=check_choosen_story(ctx.message.author, list_svr_story), timeout=60.0)
+                    temp_msg1 = await self.bot.wait_for(event="message", check=self.check_choosen_story(ctx.message.author, list_svr_story), timeout=60.0)
                     picked_story = list_svr_story[int(temp_msg1.content) - 1]
                     self.conn.cursor.execute("""DELETE FROM ows_results WHERE server_id=:sid AND story=:story""", {"sid":str(ctx.message.guild.id), "story":picked_story[3]})
                     await ctx.send("You have thrown away your **{}** Book.".format(picked_story[2]))
@@ -69,58 +97,61 @@ class OWS(commands.Cog):
                 except asyncio.TimeoutError:
                     await temp_msg.delete()
                     await xmsg.delete()
+                except commands.MissingPermissions:
+                    emb = discord.Embed(colour=discord.Colour(WHITE))
+                    emb.add_field(name="You Can't Delete Story", value="`You Don't Have Permission to do that.`")
+                    await ctx.send(embed = emb)
 
-        except Exception as exc:
-            if type(exc) == commands.BadArgument:
-                emb = discord.Embed(colour=discord.Colour(WHITE))
-                emb.set_author(name="📖 One Word Story")
-                emb.add_field(name="Commands (alias):", value="""***Start*** *-> Start The Game in the Channel.*\n***Help (h)*** *-> Help about One Word Story.*\n***OurStory (os)*** *-> Books Writen by your Server.*\n***Random (r)*** *-> Generate Random Story from Everywhere.*\n***How*** *-> The Manual to play this Game.*""", inline=False)
-                emb.set_footer(text="Example Command : g.ows how")
-                await ctx.send(embed=emb)
-            elif type(exc) == commands.MissingPermissions:
-                emb = discord.Embed(colour=discord.Colour(WHITE))
-                emb.add_field(name="You Can't Delete Story", value="`You Don't Have Permission to do that.`")
-                await ctx.send(embed = emb)
-            else:
-                print(type(exc), exc)
+        except commands.BadArgument:
+            emb = discord.Embed(colour=discord.Colour(WHITE))
+            emb.set_author(name="📖 One Word Story")
+            emb.add_field(name="Commands (alias):", value="""***Start*** *-> Start The Game in the Channel.*\n***Help (h)*** *-> Help about One Word Story.*\n***OurStory (os)*** *-> Books Writen by your Server.*\n***Random (r)*** *-> Generate Random Story from Everywhere.*\n***How*** *-> The Manual to play this Game.*""", inline=False)
+            emb.set_footer(text="Example Command : g.ows how")
+            await ctx.send(embed=emb)
 
-    @ows.error
+    @ows.error # Occure OWS error
     async def ows_error(self, ctx, error):
         if isinstance(error, commands.MissingRequiredArgument):
             emb = discord.Embed(colour=discord.Colour(WHITE))
             emb.set_author(name="~ One Word Story ~")
             emb.add_field(name="Commands (alias):", value=open("./DataPack/Help/owsh.txt", 'r').read(), inline=False)
             await ctx.send(embed=emb)
-            
+
     async def ows_gameplay(self, ctx):
-        # Initialize first Embed
+        # Initialize first
         emb = discord.Embed(colour=discord.Colour(WHITE))
-        next_value, on_play, new_emb, this_channel_id = [], True, discord.Embed(colour=discord.Colour(WHITE)), ctx.message.channel.id
+        chn_id = ctx.message.channel.id
+
+        # Attribute in-game
+        next_value = []
+        on_play = True
+        new_emb = discord.Embed(colour=discord.Colour(WHITE))
         handler_sent_emb = discord.Embed(description="***Current Progress***", colour=discord.Colour(WHITE))
         emb.add_field(name="📖 OWS (Game Started)", value="Send a single Word by different person to make a Story.")
         emb.set_footer(text="type 'THE END' to end the Game. type 'UNDO' to undo a word. type 'CANCEL' to cancel the Game.")
         this_message = await ctx.send(embed=emb)
         other_message = await ctx.send(embed=handler_sent_emb)
         lastPerson = ctx.message.author
+
         # Game started only in The Channel
         try:
             chains = 0
             while on_play:
-                new_word = await self.bot.wait_for(event="message", check=check_user_chain(this_channel_id), timeout=86400.0)
-                if new_word.content == "THE END":
+                new_word = await self.bot.wait_for(event="message", check=self.check_in_game_command(chn_id), timeout=3600.0)
+                if new_word.content == "THE END" and new_word.channel.id == chn_id:
                     lastPerson = new_word.author
                     on_play = False
-                elif new_word.content == "UNDO":
+                elif new_word.content == "UNDO" and new_word.channel.id == chn_id:
                     if len(next_value) == 0:
                         continue
                     del next_value[len(next_value) - 1]
                     handler_sent_emb = discord.Embed(description=" ".join(next_value), colour=discord.Colour(WHITE))
                     await other_message.edit(embed=handler_sent_emb)
-                elif new_word.content == "CANCEL":
+                elif new_word.content == "CANCEL" and new_word.channel.id == chn_id:
                     await ctx.send(content="*Game has been Cancelled, Ok Next Time!*")
                     await other_message.delete()
                     await this_message.delete()
-                else:
+                elif new_word.channel.id == chn_id:
                     next_value.append(new_word.content)
                     chains += 1
                     if chains >= 25:
@@ -141,9 +172,9 @@ class OWS(commands.Cog):
         else:
             # Last Response Before The Game End, Save it into Database
             this_handler_message_author_last = await ctx.send("{} please give a Title for this Story.".format(lastPerson.mention))
-            title_story = await self.bot.wait_for(event="message", check=check_last_person_ows_give_title(lastPerson))
+            title_story = await self.bot.wait_for(event="message", check=self.check_last_person_ows_give_title(lastPerson))
             await this_handler_message_author_last.delete()
-            self.conn.cursor.execute("""INSERT INTO ows_results VALUES(:svr_id, :chn_id, :title, :story)""", {"svr_id":str(title_story.guild.id), "chn_id":str(this_channel_id), "title":title_story.content, "story":" ".join(next_value)})
+            self.conn.cursor.execute("""INSERT INTO ows_results VALUES(:svr_id, :chn_id, :title, :story)""", {"svr_id":str(title_story.guild.id), "chn_id":str(chn_id), "title":title_story.content, "story":" ".join(next_value)})
             self.conn.connect.commit()
             new_emb.add_field(name="📖 {}".format(title_story.content), value=" ".join(next_value), inline=False)
             await ctx.send(embed=new_emb)
@@ -162,7 +193,8 @@ class OWS(commands.Cog):
     async def show_list_story(self, ctx):
         emb = discord.Embed(colour=discord.Colour(WHITE))
         self.conn.cursor.execute("""SELECT * FROM ows_results WHERE server_id=:sid""", {"sid":str(ctx.message.guild.id)})
-        list_of_server_stories, titles = self.conn.cursor.fetchall(), ""
+        list_of_server_stories = self.conn.cursor.fetchall()
+        titles = ""
         if list_of_server_stories is None or len(list_of_server_stories) == 0:
             emb.add_field(name="Your Server Stories", value="```There is No Stories Yet.```")
         else:
